Update betterstack-utils extension
- Refactors
- Cleanup
