+++ extensions/betterstack-utils/src/on-call-schedule.tsx
@@ -1,10 +1,9 @@
 import { Action, ActionPanel, Detail, environment, showToast, Toast } from "@raycast/api";
-import { useEffect, useState } from "react";
-import { promises as fs } from "node:fs";
-import path from "node:path";
-import { getOnCallCalendars, getCalendarEvents } from "./api";
-import { getCurrentMonthWindow, getThreeMonthWindow, getOnCallForDay, type OnCallEvent } from "./dates";
-import { buildCombinedScheduleSvg, copyImageToClipboard, svgToPng, toSvgDataUri } from "./schedule-svg";
+import { useState } from "react";
+import { getCurrentMonthWindow, getThreeMonthWindow } from "./utils/dates";
+import { buildCombinedScheduleSvg, exportSvgToClipboard, toSvgDataUri } from "./svg/schedule-svg";
+import { useOnCallData } from "./hooks/use-on-call-data";
+import { formatUserName, getCurrentOnCallUser } from "./domain/on-call-event";
 
 type TimeRange = "current-month" | "3-months";
 
@@ -14,51 +13,9 @@ const TIME_RANGE_LABELS: Record<TimeRange, string> = {
 };
 
 export default function Command() {
-  const [isLoading, setIsLoading] = useState(true);
-  const [events, setEvents] = useState<OnCallEvent[]>([]);
-  const [scheduleName, setScheduleName] = useState("");
-  const [noSchedule, setNoSchedule] = useState(false);
+  const { events, scheduleName, isLoading, noSchedule } = useOnCallData();
   const [timeRange, setTimeRange] = useState<TimeRange>("current-month");
   const [selectedUser, setSelectedUser] = useState<string>("");
-
-  useEffect(() => {
-    async function load() {
-      try {
-        const calendars = await getOnCallCalendars();
-        const primary =
-          calendars.find((c) => c.attributes.name?.toLowerCase().includes("primary")) ??
-          calendars.find((c) => c.attributes.default_calendar);
-
-        if (!primary) {
-          setNoSchedule(true);
-          setIsLoading(false);
-          return;
-        }
-
-        setScheduleName(primary.attributes.name ?? "Primary");
-
-        const { start, end } = getThreeMonthWindow();
-        const calEvents = await getCalendarEvents(primary.id, start, end);
-
-        setEvents(
-          calEvents.map((e) => ({
-            started_at: e.attributes.started_at,
-            ended_at: e.attributes.ended_at,
-            user: e.attributes.user,
-          })),
-        );
-      } catch (error) {
-        void showToast({
-          style: Toast.Style.Failure,
-          title: "Failed to load on-call schedule",
-          message: error instanceof Error ? error.message : String(error),
-        });
-      } finally {
-        setIsLoading(false);
-      }
-    }
-    void load();
-  }, []);
 
   const today = new Date();
 
@@ -72,24 +29,16 @@ export default function Command() {
     );
   }
 
-  const userNames = [...new Set(events.map((e) => `${e.user.first_name} ${e.user.last_name}`.trim()))].sort();
-  const filteredEvents = selectedUser
-    ? events.filter((e) => `${e.user.first_name} ${e.user.last_name}`.trim() === selectedUser)
-    : events;
+  const userNames = [...new Set(events.map((e) => formatUserName(e.user)))].sort();
+  const filteredEvents = selectedUser ? events.filter((e) => formatUserName(e.user) === selectedUser) : events;
 
   async function copyAsPng() {
     const toast = await showToast({ style: Toast.Style.Animated, title: "Copying to clipboard…" });
     try {
       const svg = buildCombinedScheduleSvg(filteredEvents, today, scheduleWindow, "#1F2433", false, events);
-      const svgPath = path.join(environment.supportPath, "schedule.svg");
-      const pngPath = path.join(environment.supportPath, "schedule.png");
-      await fs.writeFile(svgPath, svg);
-      await svgToPng(svgPath, pngPath);
-      await copyImageToClipboard(pngPath);
+      await exportSvgToClipboard(svg, environment.supportPath);
       toast.style = Toast.Style.Success;
       toast.title = "Schedule copied";
-      fs.unlink(svgPath).catch(() => {});
-      fs.unlink(pngPath).catch(() => {});
     } catch (error) {
       toast.style = Toast.Style.Failure;
       toast.title = "Failed to copy schedule";
@@ -100,14 +49,10 @@ export default function Command() {
   const nextTimeRange: TimeRange = timeRange === "current-month" ? "3-months" : "current-month";
   const scheduleWindow = timeRange === "current-month" ? getCurrentMonthWindow() : getThreeMonthWindow();
 
-  const currentOnCall = isLoading ? null : getOnCallForDay(today, events);
+  const currentOnCall = isLoading ? null : getCurrentOnCallUser(today, events);
   const currentlyOnCallMessage =
     timeRange === "current-month" && currentOnCall
-      ? [
-          "",
-          `**Currently on call:** ${currentOnCall.first_name} ${currentOnCall.last_name}`.trim(),
-          `**${currentOnCall.email}**`,
-        ].join("\n")
+      ? ["", `**Currently on call:** ${formatUserName(currentOnCall)}`, `**${currentOnCall.email}**`].join("\n")
       : "";
 
   const markdown = isLoading
