+++ extensions/betterstack-utils/src/utils/dates.ts
@@ -1,0 +1,31 @@
+export function startOfWeek(date: Date): Date {
+  const dayOffset = (date.getDay() + 6) % 7;
+  return new Date(date.getFullYear(), date.getMonth(), date.getDate() - dayOffset);
+}
+
+export function addDays(date: Date, days: number): Date {
+  return new Date(date.getFullYear(), date.getMonth(), date.getDate() + days);
+}
+
+export function isSameDay(a: Date, b: Date): boolean {
+  return a.getFullYear() === b.getFullYear() && a.getMonth() === b.getMonth() && a.getDate() === b.getDate();
+}
+
+export function getCurrentMonthWindow(): { start: Date; end: Date } {
+  const now = new Date();
+  const start = new Date(now.getFullYear(), now.getMonth(), 1);
+  const end = new Date(now.getFullYear(), now.getMonth() + 1, 0);
+
+  return { start, end };
+}
+
+export function getThreeMonthWindow(): { start: Date; end: Date } {
+  const now = new Date();
+  const year = now.getFullYear();
+  const month = now.getMonth();
+
+  const start = new Date(year, month - 1, 1);
+  const end = new Date(year, month + 2, 0);
+
+  return { start, end };
+}
